+++ mayan/apps/document_indexing/forms.py
@@ -1,11 +1,20 @@
-from __future__ import unicode_literals
+from __future__ import absolute_import, unicode_literals
 
 from django import forms
+from django.utils.translation import ugettext_lazy as _
 
 from common.classes import ModelAttribute
 from documents.models import Document
 
-from .models import IndexTemplateNode
+from .models import Index, IndexTemplateNode
+
+
+class IndexListForm(forms.Form):
+    indexes = forms.ModelMultipleChoiceField(
+        help_text=_('Indexes to be queued for rebuilding.'),
+        label=_('Indexes'), queryset=Index.objects.filter(enabled=True),
+        required=False, widget=forms.widgets.CheckboxSelectMultiple()
+    )
 
 
 class IndexTemplateNodeForm(forms.ModelForm):