+++ mayan/apps/document_indexing/migrations/0011_auto_20170524_0456.py
@@ -15,6 +15,11 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='indextemplatenode',
             name='expression',
-            field=models.TextField(help_text="Enter a template to render. Use Django's default templating language (https://docs.djangoproject.com/en/1.7/ref/templates/builtins/)", verbose_name='Indexing expression'),
+            field=models.TextField(
+                help_text="Enter a template to render. Use Django's default "
+                "templating language (https://docs.djangoproject.com/en/1.7/"
+                "ref/templates/builtins/)",
+                verbose_name='Indexing expression'
+            ),
         ),
     ]